% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{package}
\name{A GOF Package}
\alias{A GOF Package}
\title{A_GOF package}
\description{
When dealing with a problem regarding Goodness-of-Fit, one must first check if the distribution under the null hypothesis
is fully specified. If all parameters are known prior and need not be estimated, the standard EDF tests may be used. This package does not cover
this case, so the reader is referred to the package 'goftest' (available on CRAN). If (some) parameters must be estimated, check if the distribution
in question is listed in the table of contents. If so, the corresponding function can be used to estimate the unspecified parameters and subsequently
calculate the desired test statistic. The null hypothesis of a sample belonging to the hypothesized distribution is rejected at some level if the test statistic does not lie in the interval given in the test
output for that particular level. Some estimation functions and functions for calculating CVM and AD statistics are also documented and may be used, but are
rather supplementary to the main test functions. Mathematical derivations of the methods used in this package are mostly given in D'Agostino, Stephens (1986).
}
\references{
D'Agostino, Stephens (1986) Goodness-of-Fit Techniques

Choulakian, Stephens (2012) Goodness-of-Fit Tests for the Generalized Pareto Distribution
}
\author{
Alexander Lappe
}
