% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{A2}
\alias{A2}
\title{Anderson-Darling test statistic}
\usage{
A2(x)
}
\arguments{
\item{x}{Vector of ordered real numbers}
}
\value{
Value of test statistic
}
\description{
Computes the Anderson-Darling test statistic \deqn{A^2 = - n - \frac{1}{n}\sum_{i=1}^{n}(2i-1)(\log Z_{i}+\log (1-Z_{n+1-i})),} for values \eqn{Z_{1},\ldots ,Z_{n}} where \eqn{Z_{i} = F(x_i)}, \eqn{F} is the assumed CDF and \eqn{x} is the ordered sample.
}
\examples{
A2(c(0.4, 0.5, 0.6))
}
