% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiscWasser1.R
\name{calcDiscWasser1}
\alias{calcDiscWasser1}
\title{Compute 1-Wasserstein distance between the fitted gene count distribution and the observed gene count distribution}
\usage{
calcDiscWasser1(vdata, vpi, vqi, nmax, method = c("mean", "median"))
}
\arguments{
\item{vdata}{A numeric vector of the empirical gene count data.}

\item{vpi}{A numeric vector of the empirical probabilities.}

\item{vqi}{A numeric vector of the theoretical probabilities.}

\item{nmax}{The maximum of the biggest count in vp_i and the biggest count in vq_i.}

\item{method}{The method used to adjust the 1-Wasserstein distance metric.}
}
\value{
calcDiscWasser1 returns a list that contains
\itemize {
\item Wdist: a vector of the unadjusted 1-Wasserstein distance for each gene.
\item Wdist.adj: a vector of adjusted 1-Wasserstein distance for each gene.
\item vDj: a vector of the difference between the two quantile functions.
}
}
\description{
This function computes the 1-Wasserstein distance between two discrete distributions.
}
