% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPPabc.R
\name{calcPPabc}
\alias{calcPPabc}
\title{Compute the area between curves (abc) in the P-P plot}
\usage{
calcPPabc(P, Q)
}
\arguments{
\item{P}{A numeric vector of the empirical probabilities.}

\item{Q}{A numeric vector of the theoretical probabilities.}
}
\value{
calcPPabc returns a vector of the area between curves (abc) values for each gene using the P-P method.
}
\description{
This function computes the area between curves in the P-P plot.
             It takes the empirical and theoretical probabilities as input and calculates the area between the 45 degree line and the CDF of the theoretical distribution against the CDF of the observed distribution.
}
