% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{est_extreme}
\alias{est_extreme}
\title{Maximum Likelihood Estimator for a Gumbel distribution}
\usage{
est_extreme(x, param = c(NA, NA))
}
\arguments{
\item{x}{Vector of real numbers}

\item{param}{Vector of location \eqn{\alpha} and scale \eqn{\beta > 0} of the distribution. For the unknown parameter(s), put NA.}
}
\value{
Vector of (estimated) parameters
}
\description{
Computes Maximum Likelihood Estimators for any unknown parameters of a Gumbel distribution with CDF \deqn{F(x) = \exp\left(-\exp\left(-\frac{x-\alpha}{\beta}\right)\right),} for a sample \eqn{x}.
}
\examples{
est_extreme(c(0.5, 0.6, 1.1), c(NA, NA))
}
