% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{est_gamma}
\alias{est_gamma}
\title{Maximum Likelihood Estimator for a Gamma distribution}
\usage{
est_gamma(x, param = c(NA, NA))
}
\arguments{
\item{x}{Vector of real numbers}

\item{param}{Vector of scale \eqn{\beta > 0} and shape \eqn{m > 0} of the distribution. For the unknown parameter(s), put NA.}
}
\value{
Vector of (estimated) parameters
}
\description{
Computes Maximum Likelihood Estimators for any unknown parameters of a Gamma distribution with CDF \deqn{F(x) = \frac{1}{\beta\Gamma(m)}\left(\frac{x}{\beta}\right)^{m-1}}, for a sample \eqn{x}.
}
\examples{
est_gamma(c(0.5, 0.6, 1.1), c(NA, NA))
#' @keywords internal
}
