% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{est_gpd}
\alias{est_gpd}
\title{Maximum Likelihood Estimator for a Generalized Pareto distribution}
\usage{
est_gpd(x, param = c(NA, NA))
}
\arguments{
\item{x}{Vector of real numbers}

\item{param}{Vector of scale \eqn{a > 0} and shape \eqn{k < 1} of the distribution. For the unknown parameter(s), put NA.}
}
\value{
Vector of (estimated) parameters
}
\description{
Computes Maximum Likelihood Estimators for any unknown parameters of a Generalized Pareto distribution with CDF \deqn{F(x) = 1 - \left(1 - \frac{kx}{a}\right)^{\frac{1}{k}},} for a sample \eqn{x}.
}
\examples{
est_gpd(c(0.5, 0.6, 1.1), c(NA, NA))
}
