% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{est_pareto}
\alias{est_pareto}
\title{Maximum Likelihood Estimator for a Pareto distribution}
\usage{
est_pareto(x, param = c(NA, NA))
}
\arguments{
\item{x}{Vector of real numbers}

\item{param}{Vector of scale \eqn{x_m > 0} and shape \eqn{\alpha > 0} of the distribution. For the unknown parameter(s), put NA.}
}
\value{
Vector of (estimated) parameters
}
\description{
Computes Maximum Likelihood Estimators for any unknown parameters of a Pareto distribution with CDF \deqn{F(x) = 1 - \left(\frac{x_m}{x}\right)^\alpha,} for (\eqn{x > x_m}), for a sample \eqn{x}.
}
\examples{
est_epareto(c(0.5, 0.6, 1.1), c(NA, NA))
}
