% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{F_extreme}
\alias{F_extreme}
\title{Cumulative distribution function of a Gumbel distribution}
\usage{
F_extreme(x, location, scale)
}
\arguments{
\item{x}{Vector of real numbers}

\item{location}{location parameter \eqn{\alpha} of the distribution}

\item{scale}{scale parameter \eqn{\beta > 0} of the distribution}
}
\value{
Vector of values of the distribution function
}
\description{
Computes the CDF of a Gumbel distribution given as \deqn{F(x) = \exp\left(-\exp\left(-\frac{x-\alpha}{\beta}\right)\right).}
}
\examples{
F_extreme(0.5, 1, 1)
}
