% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{F_gpd}
\alias{F_gpd}
\title{Cumulative distribution function of a Generalized Pareto distribution}
\usage{
F_gpd(x, scale, shape)
}
\arguments{
\item{x}{Vector of real numbers}

\item{scale}{scale parameter \eqn{a} of the distribution}

\item{shape}{shape parameter \eqn{k} of the distribution}
}
\value{
Vector of values of the distribution function at \eqn{x}
}
\description{
Computes the CDF of a generalized Pareto distribution given as \deqn{F(x) = 1 - \left(1 - \frac{kx}{a}\right)^{\frac{1}{k}}.}
}
\examples{
F_extreme(c(0.5, 0.6), 1, 1)
}
