% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitDist.R
\name{FitDist}
\alias{FitDist}
\title{Fit a distribution to gene count distribution}
\usage{
FitDist(
  vdata,
  family = "average negative binomial",
  p.threshold = 1 - 10^(-5),
  common.sf = 0.7440919
)
}
\arguments{
\item{vdata}{A vector of data points - gene count data.}

\item{family}{A string specifying the model type to be used for fitting the observed distribution. By default, we fit an Average Negative Binomial distribution to the observed gene count distribution.}

\item{p.threshold}{Probability threshold used to find the big quantile of the NB or ANB fit.}

\item{common.sf}{The common spread factor parameter. By default, it is 0.7440919 which was estimated from the 3 cell line mixture data.}
}
\value{
FitDist returns a list with containing the following components:
\itemize {
\item Data: a vector of empirical gene count data.
\item FittedData: a vector of fitted gene count data.
\item vigrid: a vector of grid from 0,1,2,..., maximum of data point.
\item vpi: a vector of empirical p_i probabilities.
\item vqi: a vector of fitted q_i probabilities.
\item nmax: the maximum of the biggest count in p_i and the biggest count in q_i.
}
}
\description{
This function fits either a Negative Binonimal (NB) or Average Negative Binomial (ANB) distribution to the observed gene count distribution.
}
