% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof-package.R
\docType{package}
\name{gof-package}
\alias{gof-package}
\alias{gof}
\title{Model-diagnostics based on cumulative residuals}
\description{
Model-diagnostics based on cumulative residuals
}
\examples{
example(cumres)

}
\references{
D.Y. Lin and L.J. Wei and Z. Ying (2002) \emph{Model-Checking
Techniques Based on Cumulative Residuals}. Biometrics, Volume 58, pp 1-12.

John Q. Su and L.J. Wei (1991) \emph{A lack-of-fit test for the mean function
in a generalized linear model}. Journal. Amer. Statist. Assoc., Volume 86,
Number 414, pp 420-426.
}
\seealso{
\code{\link[timereg]{cox.aalen}} in the \code{timereg}-package for
similar GoF-methods for survival-data.
}
\author{
Klaus K. Holst <klaus@holst.it>
}
\keyword{package}
