% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gof_exp}
\alias{gof_exp}
\title{Goodness-of-fit Test for the exponential distribution}
\usage{
gof_exp(x, statistic)
}
\arguments{
\item{x}{Vector of real numbers}

\item{statistic}{Either 'cvm' or 'ad'. Specifies which of the two tests to compute}
}
\value{
Object of class gof_test
}
\description{
Computes either Cramér-von Mises or Anderson-Darling test for Goodness-of-Fit of the exponential distribution for a sample \eqn{x}. Unknown parameters are first estimated. The CDF is given by \deqn{F(x) = 1 - \exp(-\lambda x).}
}
\examples{
gof_exp(c(0.02, 0.03, 1.1), 'cvm')
}
