% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gof_extreme}
\alias{gof_extreme}
\title{Goodness-of-fit Test for the Gumbel/Extreme value distribution}
\usage{
gof_extreme(x, statistic, param = c(NA, NA))
}
\arguments{
\item{x}{Vector of real numbers}

\item{statistic}{Either 'cvm' or 'ad'. Specifies which of the two tests to compute}

\item{param}{vector of location \eqn{\alpha} and scale \eqn{\beta > 0} of the distribution. For the unknown parameter(s), put NA.}
}
\value{
Object of class gof_test
}
\description{
Computes either Cramér-von Mises or Anderson-Darling test for Goodness-of-Fit of the Gumbel/Extreme value distribution for a sample \eqn{x}. Unknown parameters are first estimated. The CDF is given by \deqn{F(x) = \exp\left(-\exp\left(-\frac{x-\alpha}{\beta}\right)\right).}
}
\examples{
gof_extremel(c(1, 2, 3), 'ad', c(NA, NA))
}
