% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gof_gamma}
\alias{gof_gamma}
\title{Goodness-of-fit Test for the Gamma distribution}
\usage{
gof_gamma(x, statistic, param = c(NA, NA))
}
\arguments{
\item{x}{Vector of real numbers}

\item{statistic}{Either 'cvm' or 'ad'. Specifies which of the two tests to compute}

\item{param}{vector of scale \eqn{\beta > 0} and shape \eqn{m > 0} of the distribution. For the unknown parameter(s), put NA.}
}
\value{
Object of class gof_test
}
\description{
Computes either Cramér-von Mises or Anderson-Darling test for Goodness-of-Fit of the Gamma distribution for a sample \eqn{x}. Unknown parameters are first estimated. The CDF is given as \deqn{F(x) = \frac{1}{\beta\Gamma(m)}\left(\frac{x}{\beta}\right)^{m-1}.}
}
\examples{
gof_gamma(c(1,2,3), 'ad', c(NA, NA))
}
