% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gof_gpd}
\alias{gof_gpd}
\title{Goodness-of-fit Test for the Generalized Pareto distribution}
\usage{
gof_gpd(x, statistic, param = c(NA, NA))
}
\arguments{
\item{x}{Vector of real numbers}

\item{statistic}{Either 'cvm' or 'ad'. Specifies which of the two tests to compute}

\item{param}{vector of scale \eqn{a > 0} and shape \eqn{k < 1} of the distribution. For the unknown parameter(s), put NA.}
}
\value{
Object of class gof_test
}
\description{
Computes either Cramér-von Mises or Anderson-Darling test for Goodness-of-Fit of the GPD for a sample \eqn{x}. Unknown parameters are first estimated. The CDF is given as \deqn{F(x) = 1 - \left(1 - \frac{kx}{a}\right)^{\frac{1}{k}}.}
}
\examples{
gof_gpd(c(1,2,3), 'ad', c(NA, NA))
}
