% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gof_normal}
\alias{gof_normal}
\title{Goodness-of-fit Test for the normal distribution}
\usage{
gof_normal(x, statistic, param = c(NA, NA))
}
\arguments{
\item{x}{Vector of real numbers}

\item{statistic}{Either 'cvm' or 'ad'. Specifies which of the two tests to compute}

\item{param}{vector of mean \eqn{\mu} and variance \eqn{\sigma^2} of the distribution. For the unknown parameter(s), put NA.}
}
\value{
Object of class gof_test
}
\description{
Computes either Cramér-von Mises or Anderson-Darling test for Goodness-of-Fit of the normal distribution for a sample \eqn{x}. Unknown parameters are first estimated. The procedure accounts for shifts of critical values due to parameter estimation.
}
\examples{
sim <- rnorm(100, 0, 1)
gof_normal(sim, 'cvm', c(NA, NA))
}
