% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gof_pareto}
\alias{gof_pareto}
\title{Goodness-of-fit Test for the Pareto distribution}
\usage{
gof_pareto(x, param)
}
\arguments{
\item{x}{Vector of real numbers}

\item{param}{scale parameter \eqn{x_m} of the distribution}

\item{statistic}{Either 'cvm' or 'ad'. Specifies which of the two tests to compute}
}
\value{
Object of class gof_test
}
\description{
Computes either Cramér-von Mises or Anderson-Darling test for Goodness-of-Fit of the Pareto distribution for a sample \eqn{x}. The CDF is given as \deqn{F(x) = 1 - \left(\frac{x_m}{x}\right)^\alpha,} (\eqn{x > x_m}). The only case that is implemented is the one of known scale \eqn{x_m} and unknown shape \eqn{\alpha}.
}
\examples{
sim <- unif(100, 1, 1)
gof_pareto(sim, 'cvm', 1)
}
