% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gof_uniform}
\alias{gof_uniform}
\title{Goodness-of-fit Test for the uniform distribution}
\usage{
gof_uniform(x)
}
\arguments{
\item{x}{Vector of real numbers}
}
\value{
Object of class gof_test
}
\description{
Computes the Cramér-von Mises test for Goodness-of-Fit of the uniform distribution for a sample \eqn{x}.
}
\examples{
sim <- runif(100, 0, 1)
gof_uniform(sim)
}
