% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{gof_weibull}
\alias{gof_weibull}
\title{Goodness-of-fit Test for the Weibull distribution}
\usage{
gof_weibull(x, statistic, param = c(NA, NA))
}
\arguments{
\item{x}{Vector of real numbers}

\item{statistic}{Either 'cvm' or 'ad'. Specifies which of the two tests to compute}

\item{param}{vector of scale \eqn{\beta > 0} and shape \eqn{m > 1} of the distribution. For the unknown parameter(s), put NA.}
}
\value{
Object of class gof_test
}
\description{
Computes either Cramér-von Mises or Anderson-Darling test for Goodness-of-Fit of the Weibull distribution for a sample \eqn{x}. Unknown parameters are first estimated. The CDF is given by \deqn{F(x) = 1 - \exp\left(\left(\frac{x}{\beta}\right)^{m}\right).}
}
\examples{
gof_weibulll(c(1,2,3), 'cvm', c(NA, NA))
}
