% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppplot.R
\name{ppplot}
\alias{ppplot}
\title{P-P plot comparing two probabilities.}
\usage{
ppplot(P, Q, prob1, prob2)
}
\arguments{
\item{P}{A numeric vector of the empirical probabilities.}

\item{Q}{A numeric vector of the theoretical probabilities.}

\item{prob1}{A character to denote the first probability distribution.}

\item{prob2}{A character to denote the first probaility distribution.}
}
\value{
ppplot returns a P-P plot using ggplot().
}
\description{
A function for generating a P-P plot to compare whether two samples come from the same distribution.
}
