% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_small_test.R
\name{qqplot_small_test}
\alias{qqplot_small_test}
\title{Q-Q plot comparing two samples with a few integer values.}
\usage{
qqplot_small_test(P, Q, sample1, sample2)
}
\arguments{
\item{P}{A numeric vector of sampled data points from the first distribution.}

\item{Q}{A numeric vector of sampled data points from the second distribution.}

\item{sample1}{A character to denote the first distribution.}

\item{sample2}{A character to denote the second distribution.}
}
\value{
qqplot_small_test returns a Q-Q plot using ggplot().
}
\description{
A function for generating a Q-Q plot to assess whether two samples with a few integer values come from the same distribution.
}
\examples{
P <- sample(0:2,1000, replace=TRUE, prob=c(1/3, 1/2, 1/6))
Q <- sample(0:1,1000, replace=TRUE, prob=c(2/3, 1/3))
qqplot_small_test(P, Q, "P", "Q")

}
\references{
{
Pan, Y., Landis, J.T., Moorad, R. et al. The Poisson distribution model fits UMI-based single-cell RNA-sequencing data.
\emph{BMC Bioinformatics 24, 256 (2023).}
\url{https://doi.org/10.1186/s12859-023-05349-2}
}
}
