% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGOF.R
\name{RunGOF}
\alias{RunGOF}
\title{Run Goodness of Fit (GOF) methods for feature selection.}
\usage{
RunGOF(
  counts,
  countsFit,
  method = c("PP", "QQ", "Wdist.mean", "Wdist.med"),
  top.n = 2000
)
}
\arguments{
\item{counts}{A raw count matrix with genes in the row and cells in the column.}

\item{countsFit}{A list generated after running FitDist().}

\item{top.n}{Number of features to select.
(default: 2000)}
}
\value{
Depending on the choice of method, RunGOF returns a list that contains
\itemize {
\item PP.ABC: a vector of ABC values for each gene using the P-P method.
\item topPP: top features selected from P-P method.
\item QQ.ABC: a vector ABC values for each gene using the Q-Q method.
\item topQQ: top features selected from Q-Q method.
\item Wdist.mn: a vector of 1-Wasserstein distance for each gene adjusted by gene mean expression.
\item topWdist.mn: top features selected from 1-Wasserstein distance adjusted by gene mean expression method.
\item Wdist.med: a vector of 1-Wasserstein distance for each gene adjusted by gene median expression.
\item topWdist.med: top features selected from 1-Wasserstein distance adjusted by gene median expression method.
}
}
\description{
Take a raw count matrix (gene in the row and cell in the column) and the NB or ANB fitted output, runs one of the 4 GOF variants for feature selection.
The output of this function is a list that contains the goodness of fit measures for all genes in the count matrix and top.n features selected based on the goodness of fit
measures. Note: both Wdist.mean and Wdist.med methods further filter the genes based on their sample and theoretical variance and only keep the genes with their sample variance
greater than their theoretical variance.
}
