% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_GOF.R
\name{theme_GOF}
\alias{theme_GOF}
\title{GOF theme ggplots}
\usage{
theme_GOF(base_size = 14, angle = 0, strip.text_size = 12)
}
\arguments{
\item{base_size}{Base font size in pts.}

\item{angle}{Angle (in [0, 45, 90]) for the axis label.}

\item{strip.text_size}{Text size in pts.}
}
\value{
theme_GOF returns a list that can be added to a ggplot object.
}
\description{
This function generates ggplot object with theme elements that are used in the plots used in the GOF package.
}
