% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{W2}
\alias{W2}
\title{Cramer-von Mises test statistic}
\usage{
W2(x)
}
\arguments{
\item{x}{Vector of ordered real numbers}
}
\value{
Value of test statistic
}
\description{
Computes the Cramér-von Mises test statistic \deqn{\sum_{i=1}^{n}\left(Z_{i}-\frac{2i-1}{2n}\right)^2 + \frac{1}{12n},} for values \eqn{Z_{1},\ldots ,Z_{n}} where \eqn{Z_{i} = F(x_i)}, \eqn{F} is the assumed CDF and \eqn{x} is the ordered sample.
}
\examples{
W2(c(0.4,0.5,0.6))
}
