% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasser1_plot.R
\name{wasser1_plot}
\alias{wasser1_plot}
\title{Make diagnostic plot for discrete 1-Wasserstein distance method}
\usage{
wasser1_plot(vigrid, vpi, vqi, vDj, Wdist, nmax, GeneName)
}
\arguments{
\item{vigrid}{A vector of grid from 0,1,2,..., maximum of data point.}

\item{vpi}{A numeric vector of the empirical probabilities.}

\item{vqi}{A numeric vector of the theoretical probabilities.}

\item{vDj}{A vector of the difference between the two quantile functions.}

\item{Wdist}{A string of the 1-Wasserstein distance metric.}

\item{nmax}{The maximum of the biggest count in vp_i and the biggest count in vq_i.}

\item{GeneName}{The name of the gene.}
}
\value{
wasser1_plot returns a diagnostic plot using ggplot().
        Left panel is a bar plot of the empirical gene count distribution and the theroretical gene count distribution.
        Right panel is a bar plot of the difference between the two distributions.
}
\description{
This function generates diagnostic plots for assessing the 1-Wasserstein distance method.
}
