\docType{package}
\name{gof-package}
\alias{gof}
\alias{gof-package}
\title{Model-diagnostics based on cumulative residuals}
\description{
  Model-diagnostics based on cumulative residuals
}
\examples{
example(cumres)
}
\author{
  Klaus K. Holst <kkho@biostat.ku.dk>
}
\references{
  D.Y. Lin and L.J. Wei and Z. Ying (2002)
  \emph{Model-Checking Techniques Based on Cumulative
  Residuals}. Biometrics, Volume 58, pp 1-12.

  John Q. Su and L.J. Wei (1991) \emph{A lack-of-fit test
  for the mean function in a generalized linear model}.
  Journal. Amer. Statist. Assoc., Volume 86, Number 414, pp
  420-426.
}
\seealso{
  \code{\link[timereg]{cox.aalen}} in the
  \code{timereg}-package for similar GoF-methods for
  survival-data.
}
\keyword{package}

