\name{surgunit}
\docType{data}
\alias{surgunit}
\title{Surgical Unit Data}
\description{
  Surgical Unit Data used in the paper by Lin et al. (2002).
  Survival time and covariates for 54 patients undergoing liver surgery.
}
\source{
Neter, J., Kutner, M. H., Nachtsheim, C. J., and Wasserman, W. (1996), Applied Linear Statistical Models, 4th edition. Chicago: Irwin}
\usage{surgunit}
\references{
  D.Y. Lin and L.J. Wei and Z. Ying (2002) \emph{Model-Checking Techniques
  Based on Cumulative Residuals}. Biometrics, Volume 58, pp 1-12.
}
